XXX This is not the final staget pp_sys.c: Improve -B and -T 8-bit handling
This removes the ASCII dependency of this code and makes it more
suitable for modern 8-bit text.

If the text up to within the final potentially multi-byte character of
the buffer is UTF-8, -T returns true, but if all the text is ASCII, it
continues looking for the ratio of controls to non-controls..

It assumes if it's text it will be Latin1, unless within the scope of
'use locale', in which case it instead assumes that locale.  Previously,
it ignored locale unless it failed ASCII text tests.