mktables: Verify Unihan files versions
Most Unicode files contain versioning information.  Prior to this patch,
mktables did not verify that the Unihan inputs were the correct version.
This is in part because we don't usually compile the Unihan properties,
and because the beta files for Unicode 7.0 had the version information
in the same syntax as the rest of the files, so I thought we only had to
wait till 7.0 came out, and we would get the protection automatically
with the code that was already there.  But the format changed between
beta and release, so this patch adds code to check for the released
format.