regexec.c: Refactor \b code to not use the FLAGS field
In one spot, \b and \B used the regnode FLAGS field to find out what
the character set type of bound they are.  This information is already
conveyed in the regnode type: BOUNDL, BOUNDU, etc.  By removing the use
of this redundant information, we can in a later commit stop storing it,
and free up the field for other uses.