Move _get_regclass_nonbitmap_data() to regcomp.c
This function, though mostly used by regexec.c and previously contained
therein, relies on the particular data structure in regcomp.c, so it
makes sense to move it adjacent to the code that generates the
structure.  Future commits will add function calls to this function that
are (currently) only in regcomp.c, and so will tie it tighter to that
file.

This just moves the code with surrounding #ifdef, adds a couple of blank
lines and adjusts the comments.  No other changes were made.