COW Static support
Cow strings use a counter behind the final "\0",
we are going to use its maximum value of 255
as an immortal counter. (using the constant COW_STATIC)

In addition to use it for B::C in order to use static cowpvs
we will also make it available for any uncompiled programs,
when a COW strings reach the 255 it will never decrease / increase.