prevent sv_dump() from being called from within sv_dump()
Note this only protects the sv_dump() wrapper call, which
may be called from debugging code which might actually be
used by sv_dump() itself. In general this doesn't produce
an infinite loop, but it could, and it certainly makes
the output less useful.

So this flag var prevents it from being called twice.
If there was a bitmap I could use instead of a full interpvar
I would.