Increase hash colision ratio and reduce hsplit calls
Previously the hash array was resized each time the hash contains
the same number of elements of its size.

With this commit, hash collision is increased to 2*SIZE, which
has two advantages: resize the hash less often, reduce the size
of the array for hashes.

As elements are now going to collide a little more, this is
going to make hash lookup a bit slower.

Note that when resizing the hash we are still using 2*Number_of_elements
(which is 4*Size_of_Hash).