Optimize File::Find performance for backup metadata building
case CPANEL-19678

Avoid recompiling the skip_pattern regex every time we see a file
if the setting is the default.  Since a regex that uses a varible
has to be recompiled with regcomp() every time it is seen we can
avoid this slowdown for the default.

Additionally Is_VMS and Is_Win32 are now constants so the code
that would enter these blocks can be avoided on platforms that
do not need it.  The Is_VMS checks were especially expensive
since they had to be run at least once per file.