Fix compilation issues with version older than 5.010
By adding more Perl versions to the Travis CI configuration
file we discovered that versions earlier 5.010 are broken.

Note that the uv.t does not raise the overflow warning
as expected by the test and we probably want to fix this in
a short future.

The test uv.t is simply not passing with 5.006 versions and
has several failures. I've currently disabled it.

Patch provided by Pali