Update minimum perl version to 5.003_07
It used to say plain 5.003, but for many years the earliest data had
been 5.003_07.  I have tested it back to there and think this is good
enough.  This is actually quite early in the 5.003 cycle.  If we wanted
to try another 5.003, it would be more likely 5.003_97 which perlhist
says was widely used, but is real close to 5.004