devel/mktodo: Add option to speed up debugging
'debug-start=' allows the caller to skip all releases above the one
specified, so that you don't have to go through them all again just to
get to one thats failing.  (The order things are done is descending.)

devel/regenerate is also made to recognize this parameter and pass it
along.