devel/scanprov: Scan for functions we didn't test
mktodo.pl tried to compile each element it knows about.  But not all
will actually compile, as some compile only when certain Configure
options are specified on certain platforms, like ones dealing with quad
math, or GLOBAL_PRIVATE_STRUCT.  But those functions are in embed.fnc
and the hdrs that get generated from that.  We can use the same
mechanism as elsewhere in scanprov to look for the first mentions of
these.