Update Net::Ping to upstream version 2.72
This retains blead customizations:
* 1a58b39af8 remove of 'use vars'
* 7bfdd8260c 500_ping_icmp.t: remove sudo code

This is mainly a unit test changes / noop changes.

Customized files are:
 t/000_load.t
 t/001_new.t
 t/010_pingecho.t
 t/500_ping_icmp.t
 t/501_ping_icmpv6.t

Files not ported to blead:
 t/020_external.t
 t/600_pod.t
 t/601_pod-coverage.t

Do not list Net::Ping customized files as mentioned by 48a109d29

Note: we could have considered doing nothing for this update,
but this would avoid updating to the CPAN version, and preserved
the existing patches.