regcomp.c: Remove memory leak
This fixes GH #18604.  There was a path through the code where a
particular SV did not get its reference count decremented.

I did an audit of the function and came up with several other
possiblities that are included in this commit.

Further, there would be leaks for some instances of finding syntax
errors in the input pattern, or when warnings are fatalized.  Those
would require mortalizing some SVs, but that is beyond the scope of this
commit.

(cherry picked from commit 5f41fa466a67b5535aa8bcf4b814f242545ac7bd)
Signed-off-by: Nicolas Rochelemagne <rochelemagne@cpanel.net>